Make Examples runnable in heartbeat / docker (#98)
This PR aims to make the examples runnable in docker by fixing a number of things. It also replaces all the examples with a simple TODOS app provided by vue.js which better demonstrate how to use playwright in a real-world situation.

As a follow-up to this PR we should add E2E tests that validate that the examples actually run. I initially wanted to do that in this PR, but I'm OK testing that elsewhere.