Support installing docker deps in docker examples (#103)
This change will upgrade any specified suites to the local build's @elastic/synthetics package. This corresponds with the change in elastic/beats@2a290a3

This also fixes some minor bugs with the docker shell scripts and makes it so that we give the container write access to the example shared volume so npm install can run. I considered copying stuff to an internal directory, but I think that's a bit much for where we are at the moment.

This can be tested by checking out this branch and:

cd examples/docker
./pull-latest.sh
./run-build-local.sh 7.10.0
verify that the npm install has run correctly in the output, maybe double-check things by removing node_modules first before running it in the example

You can run this against your own ES box by passing heartbeat options at the third step like

./run-build-local.sh 7.10.0 -E output.elasticsearch.hosts=myhost etc.