chore: update to node 16 and drop npm check (#503)
Update to the latest stable version of NodeJS 16.15.0
Follow up of HB PR - [Heartbeat] Update agent-complete node version to 16.15.0 beats#31675
Remove the dependency of NPM 6 which moves us to the lates 8 and also moves us to the new pkg lock format 2.
Update Jest deps that has Minimist <=1.2.5 is vulnerable to Prototype Pollution.
Jest also started supporting exports field for resolving modules and playwright-core did not have the generator as part of the exported library code which stopped us from requiring that module. We have a custom resolver now that patches that and resolves to old way.