Only append 'Elastic/Synthetics' to default UA
When we originally implemented
https://github.com/elastic/synthetics/issues/232 it appears we didn't
implement the third AC, allowing users to override our custom UA string
"Elastic/Synthetics" which is appended.

Some users need to fully replace the UA, this allows them to do that.
Rather than adding a boolean, this assumes that if you're overriding the
UA you want full control, which is more in line with the principle of
least surprise.