Add apiRequest parameter to journey context
This PR adds support for [API Requests](https://playwright.dev/docs/test-api-testing#sending-api-requests-from-ui-tests) in browser tests.

Note, this is not a facility for writing pure API tests, but rather a way to support
writing plain HTTP requests in service of a browser based test (as indicated in the above-linked playwright docs).

As an example, a user may want to retrieve a set of credentials from a vault based on HTTP,
or get a piece of metadata from a JSON API.

While it would be possible for users to import `requests` from `playwright-core`
and achieve a similar effect this PR is superior in that:

1. It provides an officially sanctioned way to accomplish this common pattern.
2. It automatically initializes the client for users, saving a line of boilerplate